+++ src/pystackapi/site.py
@@ -68,6 +68,10 @@ class Site(BaseClient):
         return Item(self.post(f'questions/{q_id}/suggested-edit/add', title=title, body=body,
                               tags=';'.join(tags), comment=comment, **kwargs)['items'][0])
 
+    def delete_answer(self, a_id: int, **kwargs: Any) -> None:
+        """Deletes an answer."""
+        self.post(f'answers/{a_id}/delete', **kwargs)
+
     def edit_answer(self, a_id: int, body: str, **kwargs: Any) -> Item:
         """Edit an existing answer."""
         return Item(self.post(f'answers/{a_id}/edit', body=body, **kwargs)['items'][0])