+++ tests/test_site/post/__init__.py
@@ -2,5 +2,6 @@
 
 from . import test_add
 from . import test_add_suggested_edits
+from . import test_delete
 from . import test_edit
 from . import test_post